// libs/common/src/types/roles.enum.ts
export enum RoleCode {
  ADMIN_GLOBAL = 'ADMIN_GLOBAL',
  CLIENT = 'CLIENT',
  CANDIDATE_KBS = 'CANDIDATE_KBS',
  KCA_CERTIFIED = 'KCA_CERTIFIED',
  AGENT = 'AGENT',
  ADMIN_KBS = 'ADMIN_KBS',
  ADMIN_KAMNET = 'ADMIN_KAMNET',
  ADMIN_LANDS = 'ADMIN_LANDS',
  STAFF_VERIFY = 'STAFF_VERIFY',
  STAFF_VALUATION = 'STAFF_VALUATION',
  PARTNER_GEO = 'PARTNER_GEO',
}